添加人员足迹地图绘制曲线

+++ iWorker/iWorker/人员管理/Models/ReportTrackPachM.swift
@@ -38,7 +38,7 @@ struct ReportTrackPachM: Codable {
 // MARK: - LocationList
 struct ReportLocationM: Codable {
     let reportDate, reportDateRemark, reportTime: String
-    let longitude, latitude: Int
+    let longitude, latitude: Double
     let location: String
 
     enum CodingKeys: String, CodingKey {
